TST: Finalize design for fixture based tests
TODO(rg): Document this!?
