+++ numpy/f2py/tests/conftest.py
@@ -1,5 +1,6 @@
 import pytest
 from . import util
+from numpy.testing import IS_WASM
 
 
 @pytest.fixture(scope="session")
@@ -11,6 +12,8 @@ def check_compilers():
     return checker
 
 
+@pytest.mark.slow
+@pytest.mark.skipif(IS_WASM, reason="Cannot start subprocess")
 @pytest.fixture(scope="module")
 def module_builder_factory(check_compilers):
     def build_module(spec):
@@ -32,3 +35,10 @@ def module_builder_factory(check_compilers):
             pytest.skip(f"Module build failed: {e}")
 
     return build_module
+
+
+# For generating modules from different specs
+@pytest.fixture(scope="module")
+def _mod(module_builder_factory, request):
+    spec = request.getfixturevalue(request.param)
+    return module_builder_factory(spec)
